alterações nos endpoints

+++ src/main/java/br/com/serasa/tarefa/repository/TarefaRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.serasa.tarefa.domain.Tarefa;
 
-public interface TarefaRepo extends JpaRepository<Tarefa, Integer> {
+public interface TarefaRepo extends JpaRepository<Tarefa, Long> {
 
 }
